+++ .gitignore
@@ -24,3 +24,4 @@ dist-ssr
 *.sw?
 
 *.iml
+/.angular/
