--- index.html
@@ -1,57 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <link rel="stylesheet" href="pico.min.css">
-    <link rel="stylesheet" href="pico.colors.min.css">
-    <title>Cueter</title>
-</head>
-<body class="pico-background-slate-800">
-<header class="container-fluid pico-background-slate-700">
-    <nav class="pico-color-lime-200">
-        <ul>
-            <li>
-                <h3 class="pico-color-lime-200">Cueter</h3>
-            </li>
-        </ul>
-    </nav>
-</header>
-<main class="container-fluid">
-    <div class="grid">
-        <div>
-            <form action="#" id="parse-form" class="grid">
-                <div>
-
-                    <fieldset role="group">
-                        <input type="reset" value="Clear">
-                        <input type="submit" value="Extract">
-                    </fieldset>
-                    <fieldset>
-                        <label for="parse-input">Cue file content</label>
-                        <textarea rows="30" id="parse-input" name="parse-input"></textarea>
-                    </fieldset>
-                    <fieldset>
-                        <label for="template">Extracted line template</label>
-                        <input type="text" id="template"
-                               value="{{{timestamp}}} {{{performer}}} - {{{title}}}"
-                               size="text"/>
-
-                        <p>Available tags: {{{timestamp}}} {{{performer}}} {{{title}}} {{{file}}} {{{index}}}</p>
-                    </fieldset>
-                </div>
-                <fieldset>
-                    <label for="parse-result">Extracted tracklist</label>
-                    <textarea rows="50" id="parse-result" readonly></textarea>
-                </fieldset>
-            </form>
-        </div>
-    </div>
-</main>
-<footer class="container-fluid">
-
-</footer>
-
-<script type="module" src="/src/main.ts"></script>
-</body>
-</html>
