+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cueter-2",
+  "name": "cueter",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -7,13 +7,10 @@
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "serve:ssr:cueter-2": "node dist/cueter-2/server/server.mjs"
+    "serve:ssr:cueter": "node dist/cueter/server/server.mjs"
   },
   "private": true,
   "dependencies": {
-    "@types/handlebars": "^4.0.40",
-    "@types/lodash.debounce": "^4.0.9",
-    "@types/luxon": "^3.4.2",
     "@angular/animations": "^19.0.0",
     "@angular/common": "^19.0.0",
     "@angular/compiler": "^19.0.0",
@@ -24,14 +21,17 @@
     "@angular/platform-server": "^19.0.0",
     "@angular/router": "^19.0.0",
     "@angular/ssr": "^19.0.1",
-    "express": "^4.18.2",
-    "rxjs": "~7.8.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.15.0",
+    "@types/handlebars": "^4.0.40",
+    "@types/lodash.debounce": "^4.0.9",
+    "@types/luxon": "^3.4.2",
     "chevrotain": "^11.0.3",
+    "express": "^4.18.2",
     "handlebars": "^4.7.8",
     "lodash.debounce": "^4.0.8",
-    "luxon": "^3.5.0"
+    "luxon": "^3.5.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^19.0.1",
@@ -40,12 +40,15 @@
     "@types/express": "^4.17.17",
     "@types/jasmine": "~5.1.0",
     "@types/node": "^18.18.0",
+    "autoprefixer": "^10.4.20",
     "jasmine-core": "~5.4.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
+    "postcss": "^8.4.49",
+    "tailwindcss": "^3.4.15",
     "typescript": "~5.6.2"
   }
 }
